MedicalCentre/requirements.txt
amqp==5.1.1
celery==5.2.7
Django==4.1.7
djangorestframework==3.14.0
redis==4.5.1
